requirements.txt
scikit-learn
streamlit
pandas
requests
python-dotenv